+++ trunk/sr2/src/items/ItemManager.h
@@ -34,7 +34,7 @@ namespace StoneRing
         
         Armor*       GenerateRandomGeneratedArmor     ( Item::eDropRarity rarity, int min_value, int max_value )const;
         Weapon*      GenerateRandomGeneratedWeapon    ( Item::eDropRarity rarity, double min_value, double max_value )const;
-        Item*        GenerateRandomItem      ( Item::eDropRarity rarity, int min_value, int max_value )const;
+        Item*        GenerateRandomItem               ( Item::eDropRarity rarity, double min_value, double max_value )const;
 
         Item * GetNamedItem( const std::string &name ) const;
         virtual Item * GetItem( const ItemRef & ref );
